Syncing quests now works much better - As players load in for the first time, they send their current quests to the server to get it up to speed with what they think should be active. - The server can then propagate this to other players (untested)
Still have to remove the quest stages from the active list as they complete. Also going to have to change how the server does completed quests so that it remembers what is completed. that will be for when I implement the ability to retroactively complete quets.

I also notcied that some quests do start and are active but don't get shown in the pip-boy until the next stage is sent. Idk why it happens with My Kind of Town and Come Fly With Me but they ARE actually syncing.

Also noticed that the only way to re-sync properly is to reload a save from the main menu. Will have to figure out how to detect loading a save from the game or dying or whatever.